# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ENTRAL implemented API
  version: 1.6.0
  description: >
    Gate 195 contract surface. Only paths with complete backend behavior are
    included. Canonical control-plane paths require internal Admin
    authentication. Member portfolio paths require organization membership;
    PostgreSQL row-level security resolves Human portfolio or assigned scope.
paths:
...
  /api/v1/member/organizations/{organizationId}/graph/projection:
    get:
      operationId: getMemberCanonicalGraphProjection
      summary: Get one renderer-neutral projection of the authorized canonical hierarchy.
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Versioned, permission-safe canonical graph projection.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GraphProjection" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/member/organizations/{organizationId}/graph/preferences:
    get:
      operationId: getMemberGraphViewPreferences
      summary: Read canonical defaults or the authenticated member's saved graph overrides.
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Current user-and-organization scoped graph preferences.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GraphViewPreferences" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateMemberGraphViewPreferences
      summary: Persist validated graph overrides using optimistic concurrency and idempotency.
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GraphViewPreferencesUpdateRequest" }
      responses:
        "200":
          description: Saved preferences and canonical event receipts.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GraphViewPreferencesMutationResponse" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: resetMemberGraphViewPreferences
      summary: Delete or replace a bounded saved-override scope using optimistic concurrency and idempotency.
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GraphViewPreferencesResetRequest" }
      responses:
        "200":
          description: Reset preferences and canonical event receipts.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GraphViewPreferencesMutationResponse" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/member/organizations/{organizationId}/graph/telemetry:
    post:
      operationId: recordMemberGraphRendererTelemetry
      summary: Record bounded graph performance measurements without customer payloads.
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GraphRendererTelemetryRequest" }
      responses:
        "202":
          description: Structured graph telemetry accepted by production observability.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GraphRendererTelemetryResponse" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
  responses:
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: The requested record is absent or outside the caller's scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    BadRequest:
      description: The request does not satisfy the canonical contract.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Conflict:
      description: Idempotency, scope, or expected-version conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    EntityRole:
      type: string
      enum: [ENTRAL, MARSHAL, GENERAL, COMMANDER, SOLDIER]
    EntityStatus:
      type: string
      enum: [BUILDING, ACTIVE, PAUSED, DEGRADED, RETIRED]
    HealthState:
      type: string
      enum: [HEALTHY, WATCH, DEGRADED, CRITICAL, UNKNOWN]
    GraphEntity:
      type: object
      additionalProperties: false
      required:
        - entity_id
        - organization_id
        - entity_type
        - hierarchy_level
        - authority_tier
        - authority_score
        - parent_id
        - lineage_ids
        - marshal_id
        - domain_id
        - business_id
        - stable_code
        - display_name
        - status
        - health
        - version
      properties:
        entity_id: { type: string, format: uuid }
        organization_id: { type: string, minLength: 1 }
        entity_type: { $ref: "#/components/schemas/EntityRole" }
        hierarchy_level: { type: integer, minimum: 0, maximum: 4 }
        authority_tier: { type: integer, minimum: 0, maximum: 4 }
        authority_score: { type: [number, "null"], minimum: 0, maximum: 1 }
        parent_id: { type: [string, "null"], format: uuid }
        lineage_ids:
          type: array
          minItems: 1
          maxItems: 64
          uniqueItems: true
          items: { type: string, format: uuid }
        marshal_id: { type: [string, "null"], format: uuid }
        domain_id: { type: [string, "null"], format: uuid }
        business_id: { type: [string, "null"], format: uuid }
        stable_code: { type: string, minLength: 1, maxLength: 200 }
        display_name: { type: string, minLength: 1, maxLength: 300 }
        status: { $ref: "#/components/schemas/EntityStatus" }
        health: { $ref: "#/components/schemas/HealthState" }
        version: { type: integer, minimum: 1 }
    GraphEdge:
      type: object
      additionalProperties: false
      required: [edge_id, source_id, target_id, relation_type, direction, lineage, status]
      properties:
        edge_id: { type: string, minLength: 1 }
        source_id: { type: string, format: uuid }
        target_id: { type: string, format: uuid }
        relation_type: { type: string, const: HIERARCHY }
        direction: { type: string, const: OUTBOUND }
        lineage: { type: boolean, const: true }
        status: { $ref: "#/components/schemas/EntityStatus" }
    GraphProjection:
      type: object
      additionalProperties: false
      required:
        - contract_version
        - schema_version
        - organization_id
        - projection_version
        - root_id
        - entities
        - edges
        - generated_at
        - evidence_version_reference
      properties:
        contract_version: { type: string, const: "1.0.0" }
        schema_version: { type: integer, const: 1 }
        organization_id: { type: string, minLength: 1 }
        projection_version: { type: integer, minimum: 0 }
        root_id: { type: string, format: uuid }
        entities:
          type: array
          maxItems: 100000
          items: { $ref: "#/components/schemas/GraphEntity" }
        edges:
          type: array
          maxItems: 200000
          items: { $ref: "#/components/schemas/GraphEdge" }
        generated_at: { type: string, format: date-time }
        evidence_version_reference:
          type: object
          additionalProperties: false
          required: [source, event_sequence]
          properties:
            source: { type: string, const: canonical_hierarchy }
            event_sequence: { type: integer, minimum: 0 }
    GraphSimpleSettings:
      type: object
      additionalProperties: false
      required:
        - arrangement
        - two_d_layout
        - three_d_layout
        - density
        - labels
        - connections
        - motion
        - synchronized_navigation
      properties:
        arrangement: { type: string, enum: [AUTO, SIDE_BY_SIDE, STACK, TWO_D_ONLY, THREE_D_ONLY] }
        two_d_layout: { type: string, enum: [AUTHORITY_RADIAL, HIERARCHY_TREE, DOMAIN_CLUSTERS, COMPACT_RADIAL] }
        three_d_layout: { type: string, enum: [AUTHORITY_RINGS, ELLIPTICAL_ORBITS, SPHERICAL_SHELLS, DOMAIN_CLUSTERS] }
        density: { type: string, enum: [COMPACT, BALANCED, SPACIOUS] }
        labels: { type: string, enum: [ALWAYS, RELEVANT, HOVER_OR_FOCUS, "OFF"] }
        connections: { type: string, enum: [RELEVANT, LINEAGE, DIRECT, ALL] }
        motion: { type: string, enum: [NORMAL, REDUCED, "OFF"] }
        synchronized_navigation: { type: boolean }
    GraphAdvancedSharedSettings:
      type: object
      additionalProperties: false
      required:
        - authority_band_spacing
        - authority_score_influence
        - node_scale
        - selected_node_scale
        - edge_width
        - edge_opacity
        - edge_curvature
        - label_threshold
        - label_scale
        - lineage_emphasis
        - color_mode
        - background_visible
        - grid_visible
        - legend_visible
        - animation_duration_ms
        - motion_easing
        - stable_layout_seed
        - performance_mode
        - level_of_detail
        - maximum_live_labels
        - frame_rate_cap
        - worker_usage
        - rendering_quality
      properties:
        authority_band_spacing: { type: number, minimum: 0.5, maximum: 3 }
        authority_score_influence: { type: number, minimum: 0, maximum: 0.45 }
        node_scale: { type: number, minimum: 0.25, maximum: 4 }
        selected_node_scale: { type: number, minimum: 1, maximum: 5 }
        edge_width: { type: number, minimum: 0.25, maximum: 8 }
        edge_opacity: { type: number, minimum: 0, maximum: 1 }
        edge_curvature: { type: number, minimum: 0, maximum: 1 }
        label_threshold: { type: number, minimum: 0, maximum: 1 }
        label_scale: { type: number, minimum: 0.5, maximum: 3 }
        lineage_emphasis: { type: number, minimum: 1, maximum: 5 }
        color_mode: { type: string, enum: [AUTHORITY, HEALTH, STATUS] }
        background_visible: { type: boolean }
        grid_visible: { type: boolean }
        legend_visible: { type: boolean }
        animation_duration_ms: { type: integer, minimum: 0, maximum: 5000 }
        motion_easing: { type: string, enum: [LINEAR, EASE_IN, EASE_OUT, EASE_IN_OUT] }
        stable_layout_seed: { type: string, minLength: 1, maxLength: 100 }
        performance_mode: { type: string, enum: [AUTO, QUALITY, BALANCED, PERFORMANCE] }
        level_of_detail: { type: string, enum: [AUTO, FULL, BALANCED, AGGRESSIVE] }
        maximum_live_labels: { type: integer, minimum: 0, maximum: 10000 }
        frame_rate_cap: { type: integer, enum: [30, 45, 60, 90, 120] }
        worker_usage: { type: string, enum: [AUTO, "ON", "OFF"] }
        rendering_quality: { type: string, enum: [LOW, MEDIUM, HIGH] }
    GraphAdvancedTwoDSettings:
      type: object
      additionalProperties: false
      required:
        - ring_spacing
        - tree_orientation
        - sibling_spacing
        - level_spacing
        - sector_padding
        - collision_padding
        - force_iterations
        - edge_routing
        - minimap_visible
        - grid_snapping
        - grid_size
        - fit_padding
      properties:
        ring_spacing: { type: number, minimum: 20, maximum: 2000 }
        tree_orientation: { type: string, enum: [TOP_DOWN, LEFT_RIGHT, CENTER_OUT] }
        sibling_spacing: { type: number, minimum: 4, maximum: 500 }
        level_spacing: { type: number, minimum: 10, maximum: 2000 }
        sector_padding: { type: number, minimum: 0, maximum: 0.45 }
        collision_padding: { type: number, minimum: 0, maximum: 200 }
        force_iterations: { type: integer, minimum: 0, maximum: 500 }
        edge_routing: { type: string, enum: [STRAIGHT, CURVED, ORTHOGONAL] }
        minimap_visible: { type: boolean }
        grid_snapping: { type: boolean }
        grid_size: { type: number, minimum: 1, maximum: 200 }
        fit_padding: { type: number, minimum: 0, maximum: 500 }
    GraphAdvancedThreeDSettings:
      type: object
      additionalProperties: false
      required:
        - ring_spacing
        - ellipse_eccentricity
        - orbit_tilt_degrees
        - cluster_spread
        - vertical_spread
        - depth_scale
        - collision_radius
        - camera_field_of_view
        - near_clip
        - far_clip
        - minimum_zoom
        - maximum_zoom
        - focus_distance
        - auto_orbit_enabled
        - auto_orbit_speed
        - orbit_direction
        - node_billboard
        - edge_depth_fade
        - bloom_intensity
        - lighting_intensity
        - focus_transition_ms
      properties:
        ring_spacing: { type: number, minimum: 20, maximum: 5000 }
        ellipse_eccentricity: { type: number, minimum: 0, maximum: 0.9 }
        orbit_tilt_degrees: { type: number, minimum: -85, maximum: 85 }
        cluster_spread: { type: number, minimum: 0.1, maximum: 10 }
        vertical_spread: { type: number, minimum: 0, maximum: 10 }
        depth_scale: { type: number, minimum: 0.1, maximum: 10 }
        collision_radius: { type: number, minimum: 0, maximum: 500 }
        camera_field_of_view: { type: number, minimum: 20, maximum: 100 }
        near_clip: { type: number, minimum: 0.001, maximum: 100 }
        far_clip: { type: number, minimum: 100, maximum: 100000 }
        minimum_zoom: { type: number, minimum: 0.01, maximum: 100 }
        maximum_zoom: { type: number, minimum: 0.01, maximum: 100 }
        focus_distance: { type: number, minimum: 1, maximum: 100000 }
        auto_orbit_enabled: { type: boolean }
        auto_orbit_speed: { type: number, minimum: 0, maximum: 5 }
        orbit_direction: { type: string, enum: [CLOCKWISE, COUNTERCLOCKWISE] }
        node_billboard: { type: boolean }
        edge_depth_fade: { type: boolean }
        bloom_intensity: { type: number, minimum: 0, maximum: 3 }
        lighting_intensity: { type: number, minimum: 0, maximum: 5 }
        focus_transition_ms: { type: integer, minimum: 0, maximum: 5000 }
    GraphPinnedPosition:
      type: object
      additionalProperties: false
      required: [entity_id, renderer, x, y, z]
      properties:
        entity_id: { type: string, format: uuid }
        renderer: { type: string, enum: ["2D", "3D"] }
        x: { type: number, minimum: -1000000, maximum: 1000000 }
        y: { type: number, minimum: -1000000, maximum: 1000000 }
        z: { type: [number, "null"], minimum: -1000000, maximum: 1000000 }
    GraphPreferenceSettings:
      type: object
      additionalProperties: false
      required: [simple, advanced_shared, advanced_2d, advanced_3d, pinned_positions]
      properties:
        simple: { $ref: "#/components/schemas/GraphSimpleSettings" }
        advanced_shared: { $ref: "#/components/schemas/GraphAdvancedSharedSettings" }
        advanced_2d: { $ref: "#/components/schemas/GraphAdvancedTwoDSettings" }
        advanced_3d: { $ref: "#/components/schemas/GraphAdvancedThreeDSettings" }
        pinned_positions:
          type: array
          maxItems: 5000
          items: { $ref: "#/components/schemas/GraphPinnedPosition" }
    GraphViewPreferences:
      type: object
      additionalProperties: false
      required:
        - contract_version
        - schema_version
        - preference_id
        - user_id
        - organization_id
        - source
        - settings
        - version
        - created_at
        - updated_at
        - migrated_from_schema_version
      properties:
        contract_version: { type: string, const: "1.0.0" }
        schema_version: { type: integer, const: 2 }
        preference_id: { type: [string, "null"], format: uuid }
        user_id: { type: string, format: uuid }
        organization_id: { type: string, minLength: 1 }
        source: { type: string, enum: [CANONICAL_DEFAULTS, SAVED_OVERRIDE] }
        settings: { $ref: "#/components/schemas/GraphPreferenceSettings" }
        version: { type: integer, minimum: 0 }
        created_at: { type: [string, "null"], format: date-time }
        updated_at: { type: [string, "null"], format: date-time }
        migrated_from_schema_version: { type: [integer, "null"], minimum: 1, maximum: 1 }
    GraphViewPreferencesUpdateRequest:
      type: object
      additionalProperties: false
      required: [contract_version, schema_version, expected_version, idempotency_key, settings]
      properties:
        contract_version: { type: string, const: "1.0.0" }
        schema_version: { type: integer, enum: [1, 2] }
        expected_version: { type: integer, minimum: 0 }
        idempotency_key: { type: string, minLength: 12, maxLength: 255 }
        settings: { $ref: "#/components/schemas/GraphPreferenceSettings" }
    GraphViewPreferencesResetRequest:
      type: object
      additionalProperties: false
      required: [contract_version, expected_version, idempotency_key, reset_scope]
      properties:
        contract_version: { type: string, const: "1.0.0" }
        expected_version: { type: integer, minimum: 0 }
        idempotency_key: { type: string, minLength: 12, maxLength: 255 }
        reset_scope: { type: string, enum: [ALL, SHARED, VIEW_2D, VIEW_3D, PINNED_POSITIONS] }
    GraphViewPreferencesMutationResponse:
      type: object
      additionalProperties: false
      required: [preferences, idempotent_replay, event_ids]
      properties:
        preferences: { $ref: "#/components/schemas/GraphViewPreferences" }
        idempotent_replay: { type: boolean }
        event_ids:
          type: array
          uniqueItems: true
          items: { type: string, format: uuid }
    GraphRendererTelemetryRequest:
      type: object
      additionalProperties: false
      required:
        - contract_version
        - schema_version
        - telemetry_id
        - observed_at
        - renderer
        - layout_pattern
        - projection_id
        - projection_version
        - node_count
        - edge_count
        - settings_version
        - layout_time_ms
        - render_time_ms
        - sample_window_ms
        - frame_rate_fps
        - dropped_frame_rate_ratio
        - error_code
      properties:
        contract_version: { type: string, const: "1.0.0" }
        schema_version: { type: integer, const: 1 }
        telemetry_id: { type: string, format: uuid }
        observed_at: { type: string, format: date-time }
        renderer: { type: string, enum: ["2D", "3D"] }
        layout_pattern:
          type: string
          enum:
            - AUTHORITY_RADIAL
            - HIERARCHY_TREE
            - DOMAIN_CLUSTERS
            - COMPACT_RADIAL
            - AUTHORITY_RINGS
            - ELLIPTICAL_ORBITS
            - SPHERICAL_SHELLS
        projection_id: { type: string, format: uuid }
        projection_version: { type: integer, minimum: 0 }
        node_count: { type: integer, minimum: 0, maximum: 100000 }
        edge_count: { type: integer, minimum: 0, maximum: 200000 }
        settings_version: { type: integer, minimum: 0 }
        layout_time_ms: { type: number, minimum: 0, maximum: 600000 }
        render_time_ms: { type: number, minimum: 0, maximum: 600000 }
        sample_window_ms: { type: number, minimum: 1, maximum: 600000 }
        frame_rate_fps: { type: number, minimum: 0, maximum: 1000 }
        dropped_frame_rate_ratio: { type: number, minimum: 0, maximum: 1 }
        error_code:
          type: string
          enum:
            - NONE
            - GRAPH_LAYOUT_FAILURE
            - GRAPH_RENDERER_FAILURE
            - GRAPH_WEBGL_CONTEXT_LOST
            - GRAPH_WORKER_FAILURE
            - GRAPH_EXPORT_FAILURE
            - GRAPH_PERFORMANCE_DEGRADED
    GraphRendererTelemetryResponse:
      type: object
      additionalProperties: false
      required:
        - contract_version
        - schema_version
        - telemetry_id
        - organization_id
        - accepted
        - recorded_at
      properties:
        contract_version: { type: string, const: "1.0.0" }
        schema_version: { type: integer, const: 1 }
        telemetry_id: { type: string, format: uuid }
        organization_id: { type: string, minLength: 1 }
        accepted: { type: boolean, const: true }
        recorded_at: { type: string, format: date-time }
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, message]
      properties:
        code: { type: string }
        error: { type: string }
        message: { type: string }
        requestId: { type: string }